+++ v1-0-0/src/components/ServerConnection.tsx
@@ -1,29 +1,13 @@
 "use client";
 
 import { useState } from "react";
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardHeader,
-  CardTitle,
-} from "./ui/card";
+import { Card } from "./ui/card";
 import { Button } from "./ui/button";
-import { Input } from "./ui/input";
 import { Plus, Server, Database } from "lucide-react";
 import { ServerWithName } from "@/hooks/useAppState";
-
-interface ServerFormData {
-  name: string;
-  type: "stdio" | "http";
-  command?: string;
-  args?: string[];
-  url?: string;
-  headers?: Record<string, string>;
-  env?: Record<string, string>;
-  useOAuth?: boolean;
-  oauthScopes?: string[];
-}
+import { ServerConnectionCard } from "./ServerConnectionCard";
+import { AddServerModal } from "./AddServerModal";
+import { ServerFormData } from "@/lib/types";
 
 interface ServerConnectionProps {
   connectedServerConfigs: Record<string, ServerWithName>;
@@ -31,9 +15,6 @@ interface ServerConnectionProps {
   onDisconnect: (serverName: string) => void;
   onReconnect: (serverName: string) => void;
 }
-
-// Import the new card component
-import { ServerConnectionCard } from "./ServerConnectionCard";
 
 export function ServerConnection({
   connectedServerConfigs,
@@ -44,43 +25,6 @@ export function ServerConnection({
   const [isAddingServer, setIsAddingServer] = useState(false);
   const [searchTerm, setSearchTerm] = useState("");
   const [filterType, setFilterType] = useState<"all" | "stdio" | "http">("all");
-  const [serverFormData, setServerFormData] = useState<ServerFormData>({
-    name: "",
-    type: "stdio",
-    command: "",
-    args: [],
-    url: "",
-    headers: {},
-    env: {},
-    useOAuth: true,
-    oauthScopes: ["mcp:*"],
-  });
-
-  const handleSubmit = (e: React.FormEvent) => {
-    e.preventDefault();
-    if (serverFormData.name) {
-      onConnect(serverFormData);
-      setServerFormData({
-        name: "",
-        type: "stdio",
-        command: "",
-        args: [],
-        url: "",
-        headers: {},
-        env: {},
-        useOAuth: false,
-        oauthScopes: ["mcp:*"],
-      });
-      setIsAddingServer(false);
-    }
-  };
-
-  const handleArgsChange = (value: string) => {
-    setServerFormData((prev) => ({
-      ...prev,
-      args: value.split(" ").filter((arg) => arg.trim()),
-    }));
-  };
 
   // Filter and search servers
   const filteredServers = Object.entries(connectedServerConfigs).filter(
@@ -161,165 +105,12 @@ export function ServerConnection({
         </Card>
       )}
 
-      {/* Add Server Modal/Form */}
-      {isAddingServer && (
-        <Card className="border-2 border-dashed">
-          <CardHeader>
-            <CardTitle>Add New Server</CardTitle>
-            <CardDescription>
-              Configure a new MCP server connection
-            </CardDescription>
-          </CardHeader>
-          <CardContent>
-            <form onSubmit={handleSubmit} className="space-y-4">
-              <div>
-                <label className="block text-sm font-medium mb-1">
-                  Server Name
-                </label>
-                <Input
-                  value={serverFormData.name}
-                  onChange={(e) =>
-                    setServerFormData((prev) => ({
-                      ...prev,
-                      name: e.target.value,
-                    }))
-                  }
-                  placeholder="my-mcp-server"
-                  required
-                />
-              </div>
-
-              <div>
-                <label className="block text-sm font-medium mb-1">Type</label>
-                <select
-                  value={serverFormData.type}
-                  onChange={(e) =>
-                    setServerFormData((prev) => ({
-                      ...prev,
-                      type: e.target.value as "stdio" | "http",
-                    }))
-                  }
-                  className="w-full p-2 border rounded-md bg-background cursor-pointer"
-                >
-                  <option value="stdio">STDIO</option>
-                  <option value="http">HTTP</option>
-                </select>
-              </div>
-
-              {serverFormData.type === "stdio" ? (
-                <>
-                  <div>
-                    <label className="block text-sm font-medium mb-1">
-                      Command
-                    </label>
-                    <Input
-                      value={serverFormData.command}
-                      onChange={(e) =>
-                        setServerFormData((prev) => ({
-                          ...prev,
-                          command: e.target.value,
-                        }))
-                      }
-                      placeholder="npx"
-                      required
-                    />
-                  </div>
-                  <div>
-                    <label className="block text-sm font-medium mb-1">
-                      Arguments
-                    </label>
-                    <Input
-                      value={serverFormData.args?.join(" ") || ""}
-                      onChange={(e) => handleArgsChange(e.target.value)}
-                      placeholder="-y @modelcontextprotocol/server-filesystem /path/to/directory"
-                    />
-                  </div>
-                </>
-              ) : (
-                <>
-                  <div>
-                    <label className="block text-sm font-medium mb-1">
-                      URL
-                    </label>
-                    <Input
-                      value={serverFormData.url}
-                      onChange={(e) =>
-                        setServerFormData((prev) => ({
-                          ...prev,
-                          url: e.target.value,
-                        }))
-                      }
-                      placeholder="http://localhost:8080/mcp"
-                      required
-                    />
-                  </div>
-
-                  <div className="space-y-3 p-4 border rounded-lg bg-muted/50">
-                    <div className="flex items-center space-x-2">
-                      <input
-                        type="checkbox"
-                        id="useOAuth"
-                        checked={serverFormData.useOAuth}
-                        onChange={(e) =>
-                          setServerFormData((prev) => ({
-                            ...prev,
-                            useOAuth: e.target.checked,
-                          }))
-                        }
-                        className="w-4 h-4 cursor-pointer"
-                      />
-                      <label
-                        htmlFor="useOAuth"
-                        className="text-sm font-medium cursor-pointer"
-                      >
-                        Use OAuth 2.1 Authentication
-                      </label>
-                    </div>
-
-                    {serverFormData.useOAuth && (
-                      <div>
-                        <label className="block text-sm font-medium mb-1">
-                          OAuth Scopes
-                        </label>
-                        <Input
-                          value={serverFormData.oauthScopes?.join(" ") || ""}
-                          onChange={(e) =>
-                            setServerFormData((prev) => ({
-                              ...prev,
-                              oauthScopes: e.target.value
-                                .split(" ")
-                                .filter((s) => s.trim()),
-                            }))
-                          }
-                          placeholder="mcp:* mcp:tools mcp:resources"
-                        />
-                        <p className="text-xs text-muted-foreground mt-1">
-                          Space-separated OAuth scopes. Use &apos;mcp:*&apos;
-                          for full access.
-                        </p>
-                      </div>
-                    )}
-                  </div>
-                </>
-              )}
-
-              <div className="flex gap-2 pt-4">
-                <Button type="submit" className="cursor-pointer">
-                  Connect Server
-                </Button>
-                <Button
-                  type="button"
-                  variant="outline"
-                  onClick={() => setIsAddingServer(false)}
-                  className="cursor-pointer"
-                >
-                  Cancel
-                </Button>
-              </div>
-            </form>
-          </CardContent>
-        </Card>
-      )}
+      {/* Add Server Modal */}
+      <AddServerModal
+        isOpen={isAddingServer}
+        onClose={() => setIsAddingServer(false)}
+        onConnect={onConnect}
+      />
     </div>
   );
 }
